Use brmo-init-util
